requirements.txt
psycopg2-binary==2.9.1
python-dotenv==0.19.2
types-psycopg2==2.9.0
